chore: move EGL bindings to a separate file
